+++ R/tar_resources.R
@@ -76,8 +76,9 @@
 #' @param network Output of function `tar_resources_network()`.
 #'   Settings to configure how to handle unreliable network connections
 #'   in the case of uploading, downloading, and checking data
-#'   in situations that rely on HTTP/HTTPS requests. Examples include
-#'   retries and timeouts for internal storage management operations for
+#'   in situations that rely on network file systems or HTTP/HTTPS requests.
+#'   Examples include retries and timeouts for internal storage management
+#'   operations for `storage = "worker"` (on network file systems),
 #'   `format = "url"`, `repository = "aws"`, and
 #'   `repository = "gcp"`. These settings do not
 #'   apply to actions you take in the custom R command of the target.