+++ R/class_crew.R
@@ -125,12 +125,16 @@ crew_class <- R6::R6Class(
       data <- self$exports$common
       data$target <- target
       globals <- self$exports$globals
+      resources <- target$settings$resources$crew
       self$controller$push(
         command = command,
         data = data,
         globals = globals,
         substitute = FALSE,
-        name = target_get_name(target)
+        name = target_get_name(target),
+        controller = resources$controller,
+        scale = resources$scale %|||% TRUE,
+        seconds_timeout = resources$seconds_timeout
       )
     },
     run_main = function(target) {