+++ NEWS.md
@@ -8,6 +8,7 @@
 * Downsize the `tar_runtime` object.
 * Remove the 100 Kb file size cutoff for determining whether to trust the file timestamp or recompute the hash when checking if a file is up to date (#1062). Instate the `"file_fast"` format and the `trust_object_timestamps` option in `tar_option_set()` as safer alternatives.
 * Consolidate store constructors.
+* Allow `crew` controller groups (#1065, @mglev1n).
 
 ## Speedups