+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 1.0.0.9001 (development)
+# targets 1.0.0.9002 (development)
 
 ## Bug fixes